COMPLETE CRYPTOGRAPHY
Completely replaced usages of the "Fernet" library with that of the Secure Hash Algorithm 256 ("SHA256"). I got the idea to use SHA256 from a classmate, and I found it to be much easier than Fernet. The client now encrypts their password attempt before sending it.

This cryptography is unsafe! Of course, compared to REAL cryptography, this is just child's play for the purposes of submitting a homework assignment. Therefore, dear reader, please do not send your bank details over the server.