Finished merging
Finished merging branches "Drew" and "Benjamin" because merging was unfinished in previous commit.

The infinite loop stops when Client's "cmd" is sent. To prevent every possible infinite loop, the client must send SOMETHING, no matter what, even if it's no big deal. ¿What if we add an "ERROR" command, but don't include it in the HELP list? When Server receives ERROR, it should reply, "Try again."

And no, a "continue" statement would not work. The problem is that Client hangs on the following line, but receives no answer: ```data = client.recv(SIZE).decode(FORMAT)```