FAILED
Attempted to squash infinite looping by forcing Clients to send and receive commands every time one is typed, even if it has to be the ERROR command. It works the first time, but not if you immediately do it again. Changed some try-except blocks to this, but may need to change them back if this doesn't pan out.

Added Print Red function "prRed" for error print statements. Deemed necessary because "logging.error" would wonkily de-synchronize with print statements